fix: add conftest for integration tests to set up cache directory
Created tests/integration/conftest.py with autouse fixture to properly
configure cache directory for all integration tests. This ensures the
cache directory exists before tests try to use it.

Previously, only tests in test_cache_integration.py had this fixture,
causing other integration tests to fail with FileNotFoundError when
trying to write cache metadata files.

All 30 integration tests now pass (excluding GCS test which requires credentials).
